+++ src/renderer/RecorderPreview.tsx
@@ -52,7 +52,9 @@ const RecorderPreview: React.FC = () => {
         border: '2px solid black',
         boxSizing: 'border-box',
       }}
-    />
+    >
+      Preview...
+    </Box>
   );
 };